requirements.txt
requests
python-dotenv
Pillow
kagglehub
pandas        # <— For CSV reading
google-generativeai
google-cloud-storage
pandas
